+++ seed/ruby-sdk/basic-auth-environment-variables/lib/gemconfig.rb
@@ -1,0 +1,14 @@
+# frozen_string_literal: true
+
+module SeedBasicAuthEnvironmentVariablesClient
+  module Gemconfig
+    VERSION = ""
+    AUTHORS = [""].freeze
+    EMAIL = ""
+    SUMMARY = ""
+    DESCRIPTION = ""
+    HOMEPAGE = "https://github.com/basic-auth-environment-variables/fern"
+    SOURCE_CODE_URI = "https://github.com/basic-auth-environment-variables/fern"
+    CHANGELOG_URI = "https://github.com/basic-auth-environment-variables/fern/blob/master/CHANGELOG.md"
+  end
+end