+++ seed/ruby-sdk/basic-auth-environment-variables/test/test_fern_basic_auth_environment_variables.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+require_relative "test_helper"
+require "fern_basic_auth_environment_variables"
+
+# Basic SeedBasicAuthEnvironmentVariablesClient tests
+class TestSeedBasicAuthEnvironmentVariablesClient < Minitest::Test
+  def test_function
+    # SeedBasicAuthEnvironmentVariablesClient::Client.new
+  end
+end